+++ src/extension.ts
@@ -6,43 +6,95 @@ import {
   ServerOptions,
 } from "vscode-languageclient/node";
 import { RideConfiguration } from "./configuration";
+import { StatemachineVisualization } from "./stateMachineVisualization";
 
 const execFile = util.promisify(require("child_process").execFile);
 
 let languageClient: LanguageClient;
 
+interface MermaidDiagramNotification {
+  uri: string;
+  diagram: string;
+}
+
 export async function activate(context: vscode.ExtensionContext) {
-  let configuration = new RideConfiguration(
+  const configuration = new RideConfiguration(
     vscode.workspace.getConfiguration()
   );
 
+  const { rideSupportsLsp, rideSupportsVisualization } =
+    await checkRideFeatures(configuration);
+
+  if (!rideSupportsLsp) {
+    vscode.window.showWarningMessage(
+      "Ride does not support lsp. Please upgrade."
+    );
+    return;
+  }
+
+  const languageClient = initializeLanguageClient(configuration);
+  await languageClient.start();
+  context.subscriptions.push(languageClient);
+  context.subscriptions.push(
+    vscode.debug.registerDebugAdapterDescriptorFactory(
+      "lf",
+      new DebugAdapterExecutableFactory(configuration)
+    )
+  );
+
+  if (!rideSupportsVisualization) {
+    vscode.window.showWarningMessage(
+      "Ride does not support lf state machine visualization. Please upgrade."
+    );
+    return;
+  }
+
+  const diagramStore: { [uri: string]: string } = {};
+  setUpVisualizationSupport(context, languageClient, diagramStore);
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
+      "extension.openStateMachineVisualization",
+      () => {
+        StatemachineVisualization.createWebview(context);
+      }
+    )
+  );
+
+  StatemachineVisualization.createWebview(context);
+}
+
+export async function deactivate() {
+  if (languageClient) {
+    await languageClient.stop();
+  }
+}
+
+async function checkRideFeatures(configuration: RideConfiguration) {
   let rideSupportsLsp = false;
+  let rideSupportsVisualization = false;
+
   try {
     const versionResult = await execFile(
       configuration.ridePath,
       ["--version"],
       {}
     );
-    const version: string = versionResult.stdout;
-    const versionParts = version.split(".").map((v) => parseInt(v));
-    if (
-      versionParts.length == 3 &&
-      (versionParts[0] > 0 || (versionParts[0] == 0 && versionParts[1] > 8))
-    ) {
-      rideSupportsLsp = true;
-    }
+
+    const semverSatisfies = require("semver/functions/satisfies");
+    rideSupportsLsp = semverSatisfies(versionResult.stdout, ">=0.9.0");
+    rideSupportsVisualization = semverSatisfies(
+      versionResult.stdout,
+      ">=1.0.0"
+    );
   } catch (e) {
     vscode.window.showWarningMessage("Ride executable not found!");
-    return;
   }
 
-  if (!rideSupportsLsp) {
-    vscode.window.showWarningMessage(
-      "Ride does not support lsp, consider upgrading to the latest ride version to get language server support for lf files."
-    );
-    return;
-  }
+  return { rideSupportsLsp, rideSupportsVisualization };
+}
 
+function initializeLanguageClient(configuration) {
   const serverOptions: ServerOptions = {
     command: configuration.ridePath,
     args: ["language-server"],
@@ -58,28 +110,40 @@ export async function activate(context: vscode.ExtensionContext) {
     ],
   };
 
-  // Create the language client and start the client.
-  languageClient = new LanguageClient(
+  return new LanguageClient(
     "rideLanguageClient",
     "RIDE Language Server",
     serverOptions,
     clientOptions
   );
-
-  languageClient.start();
-
-  context.subscriptions.push(
-    vscode.debug.registerDebugAdapterDescriptorFactory(
-      "lf",
-      new DebugAdapterExecutableFactory(configuration)
-    )
-  );
 }
 
-export async function deactivate() {
-  if (languageClient) {
-    await languageClient.stop();
-  }
+function setUpVisualizationSupport(context, languageClient, diagramStore) {
+  languageClient.onNotification(
+    "lfls/mermaidVisualization",
+    (notification: MermaidDiagramNotification) => {
+      const { uri, diagram } = notification;
+      diagramStore[uri] = diagram;
+      StatemachineVisualization.currentPanel?.updateStatemachineVisualization(
+        diagram
+      );
+    }
+  );
+
+  let disposable = vscode.window.onDidChangeActiveTextEditor((editor) => {
+    if (!editor) return;
+    const { fileName, languageId } = editor.document;
+
+    if (fileName.endsWith(".lf") || languageId === "lf") {
+      const diagram = diagramStore[`file://${editor.document.fileName}`];
+      if (diagram) {
+        StatemachineVisualization.currentPanel?.updateStatemachineVisualization(
+          diagram
+        );
+      }
+    }
+  });
+  context.subscriptions.push(disposable);
 }
 
 class DebugAdapterExecutableFactory