+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Rename `Franka Emika` to `Franka Robotics`.
 * Add syntax highlighting for state interfaces, dynamic links and anonymous states.
 * Integrate LF Debugger. (PSWH-327)
+* Visualize the LF state machine in a separate tab.
 
 ## 0.2.0 - 2023-07-27